Move contract related types/interfaces to the contracts utils file
